math notation to webgl layout
